+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 - fix: include value in field id for checkbox and radio elements so the focus is working when clicking on them
 - fix: replace field's null value with empty string in controlled mode
 - fix: fix checked state of radio/checkbox with empty string as value and context value is null
+- fix: update fields to validate instead of replacing them when setting values
 
 ## v5.6.0 (2025-01-09)